add commitment params struct

+++ src/consts.rs
@@ -1,8 +1,8 @@
 pub const Q: u64 = 18446744073709551557; // previous_prime(2**64)
-pub const D: u64 = 128;
+pub const D: usize = 128;
 pub const L: u64 = 64;
-pub const KAPPA: u64 = 20;
-pub const LAMBDA: u64 = 24;
-pub const ALPHA: u64 = 24;
+pub const KAPPA: usize = 20;
+pub const LAMBDA: usize = 24;
+pub const ALPHA: usize = 24;
 pub const MU: u64 = 127;
